Update tests to support changes to PublishFiles
